TestUnitaire

+++ src/main/java/com/spring/kaddem/entities/DetailEquipe.java
@@ -19,7 +19,7 @@ public class DetailEquipe implements Serializable {
     private static final long serialVersionUID = 1L;
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long idDetailEquipe;
+    private int idDetailEquipe;
     private Integer salle;
     private String thematique;
     @OneToOne(mappedBy = "detailEquipe")
